- einbinden des Airship und dem Markt aus dem GameState - Airship Preis = WARE_SELL_RATE * Market Preis - float mit double ersetzt - Für das TableModel ein Enum Character(PLAYER, OPPONENT) erstellt, um Markt.buyItem/sellItem zu verwenden können

+++ src/common/Constants.java
@@ -40,8 +40,8 @@ public class Constants {
     public static final int HARBOR_CIRCLE_DIAMETER = 20;
     public static final String HARBOR_ICON = "resources/panels/img/anchor.png";
     
-    public static final float WARE_STANDARD_AMOUNT = 100;
-    public static final float WARE_MAX_INFLATION_FACTOR = 5;
+    public static final double WARE_STANDARD_AMOUNT = 100;
+    public static final double WARE_MAX_INFLATION_FACTOR = 5;
     public static final double WARE_SELL_RATIO = 0.8;
