+++ src/model/economy/ware/Food.java
@@ -4,7 +4,7 @@ import model.economy.Ware;
 
 public class Food extends Ware {
 
-    public Food(String name, int value, int weight, float price) {
+    public Food(String name, int value, int weight, double price) {
         super(name, value, weight, price);
     }
